+++ src/main/java/handler/FileRequestHandler.java
@@ -1,7 +1,7 @@
 package handler;
 
-import http.HttpRequest;
-import http.HttpResponse;
+import http.request.HttpRequest;
+import http.response.HttpResponse;
 import http.enums.HttpMethod;
 import http.enums.HttpStatus;
 import http.enums.MimeType;
