+++ src/main/java/handler/RequestHandler.java
@@ -1,7 +1,7 @@
 package handler;
 
-import http.HttpRequest;
-import http.HttpResponse;
+import http.request.HttpRequest;
+import http.response.HttpResponse;
 
 public interface RequestHandler {
     boolean canHandle(HttpRequest httpRequest);
