+++ src/main/java/http/request/HttpRequestResolver.java
@@ -1,17 +1,12 @@
-package http;
+package http.request;
 
-import http.enums.HttpMethod;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.List;
 
 public class HttpRequestResolver {
     private static final Logger logger = LoggerFactory.getLogger(HttpRequestResolver.class);
