+++ src/main/java/http/response/HttpResponse.java
@@ -1,4 +1,4 @@
-package http;
+package http.response;
 
 import http.enums.HttpStatus;
