+++ src/main/java/webserver/WebServlet.java
@@ -3,10 +3,10 @@ package webserver;
 import exception.NotFoundRequestHandlerException;
 import handler.FileRequestHandler;
 import handler.mapping.RequestHandlerMapping;
-import http.HttpRequestResolver;
-import http.HttpResponse;
-import http.HttpResponseResolver;
-import http.HttpRequest;
+import http.request.HttpRequestResolver;
+import http.response.HttpResponse;
+import http.response.HttpResponseResolver;
+import http.request.HttpRequest;
 import http.enums.HttpStatus;
 import http.enums.MimeType;
 import org.slf4j.Logger;
